Replace batch file installer with InnoSetup installer
The batch files would not work in Windows 10 due to the way cmd.exe handles the working directory of a batch file when invoked with UAC. In Windows 7, it would set the working directory to the folder containing the batch file and xcopy deployment worked. In Windows 10, it's set to System32 and the xcopy deployment tried to install the contents of System32 to the Program Files folder. :frowning: