Fix UCCom not working at all anymore
IsOpen was set too late in the Open() function. It caused the async read method to return immediately.

If the serial port was bad, Main() didn't explicitly close it, which resulted in a really weird GUI state.

On the bright side, opening/closing the serial connection in rapid succession no longer seems to crash crtcpl.